chore: moved pilot endpoints to routes package
